requirements.txt
pandas==1.3.5
psycopg2-binary==2.9.3
python-dotenv
